requirements.txt
numpy>=1.26
pandas>=2.2
scipy>=1.13
statsmodels>=0.14
geopandas>=1.0
shapely>=2.0
fiona>=1.9
pyproj>=3.6
rtree>=1.2
osmnx>=1.9
networkx>=3.2
folium>=0.16
branca>=0.7
matplotlib>=3.9
seaborn>=0.13
contextily>=1.6
requests>=2.32
pydantic>=2.8
PyYAML>=6.0
tqdm>=4.66
joblib>=1.4
numba>=0.60
scikit-learn>=1.5
geopy>=2.4
haversine>=2.8
cachetools>=5.3
googlemaps>=4.10
openrouteservice>=2.3
pyowm>=3.3
tenacity>=9.0
ratelimit>=2.2
mapbox>=0.18
python-dateutil>=2.9
